Изменение типа идентификатора пользователя. Изменение интерфейса.

+++ CharApplication.Dbl/Models/DbFile.cs
@@ -26,7 +26,7 @@ namespace ChatApplication.Dbl.Models
         /// <summary>
         /// Автор
         /// </summary>
-        public long AuthorId { get; set; }
+        public int AuthorId { get; set; }
         /// <summary>
         /// Создано
         /// </summary>
